[compiler] add vite compiler plugin
HMR does not work for all the files.
Had to add custom logic.
