+++ examples/counter/vite.config.ts
@@ -1,8 +1,7 @@
 import { resolve } from 'path';
 import Inspect from 'vite-plugin-inspect';
 import { defineConfig } from 'vitest/config';
-import { JayRollupConfig, jayRuntime } from 'rollup-plugin-jay';
-import { ModuleNode, ViteDevServer } from 'vite';
+import { JayRollupConfig, jayRuntime } from 'vite-plugin-jay';
 
 const root = resolve(__dirname);
 const jayOptions: JayRollupConfig = {
@@ -15,28 +14,7 @@ export default defineConfig(({ mode }) => {
         mode === 'production' ? [] : ['jay-component', 'jay-reactive', 'jay-runtime', 'jay-secure'];
 
     return {
-        plugins: [
-            Inspect(),
-            {
-                enforce: 'pre',
-                ...jayRuntime(jayOptions),
-                handleHotUpdate({
-                    file,
-                    modules,
-                    server,
-                }: {
-                    file: string;
-                    modules: ModuleNode[];
-                    server: ViteDevServer;
-                }): ModuleNode[] {
-                    if (modules.length === 0 && file.endsWith('.jay-html')) {
-                        const tsFile = server.moduleGraph.getModuleById(`${file}.ts`);
-                        return tsFile ? [tsFile] : [];
-                    }
-                    return modules;
-                },
-            },
-        ],
+        plugins: [Inspect(), jayRuntime(jayOptions)],
         worker: {
             rollupOptions: {
                 external,
