+++ examples/mini-benchmark/package.json
@@ -40,14 +40,10 @@
     "jay-dev-environment": "workspace:^",
     "npm-run-all2": "^6.1.1",
     "rimraf": "^5.0.5",
-    "rollup": "^4.6.1",
-    "rollup-plugin-copy": "^3.5.0",
-    "rollup-plugin-jay": "workspace:^",
-    "rollup-plugin-terser": "^7.0.2",
-    "rollup-plugin-typescript2": "^0.36.0",
     "tslib": "^2.6.2",
     "typescript": "^5.3.2",
     "vite": "^5.0.5",
-    "vite-plugin-inspect": "^0.8.1"
+    "vite-plugin-inspect": "^0.8.1",
+    "vite-plugin-jay": "workspace:^"
   }
 }
