+++ package.json
@@ -2,7 +2,7 @@
   "name": "jay",
   "scripts": {
     "build": "wsrun -te --report build",
-    "build:watch": "wsrun -p 'jay-@(cli|compiler|component|json-patch|list-compare|reactive|runtime|secure|serialization)' -p 'rollup-plugin-jay' -m --prefix build:watch",
+    "build:watch": "wsrun -p 'jay-@(cli|compiler|component|json-patch|list-compare|reactive|runtime|secure|serialization)' -p '@(rollup|vite)-plugin-jay' -m --prefix build:watch",
     "clean": "wsrun clean",
     "clean:node_modules": "rimraf node_modules && rimraf packages/*/node_modules && rimraf examples/*/node_modules && rimraf exploration/*/node_modules",
     "clean:node_modules:with-cache": "npm run clean:node_modules && yarn cache clean",
