+++ packages/rollup-plugin/lib/files.ts
@@ -1,9 +1,9 @@
 import path from 'node:path';
 import { PluginContext } from 'rollup';
-import { JAY_DTS_EXTENSION, JAY_EXTENSION } from '../../compiler/lib/core/constants.ts';
+import { JAY_DTS_EXTENSION, JAY_EXTENSION } from '../../compiler/lib/core/constants';
 import { mkdir, readFile } from 'node:fs/promises';
 import { writeFile } from 'fs/promises';
-import { JayPluginContext } from './jay-plugin-context.ts';
+import { JayPluginContext } from './jay-plugin-context';
 
 export function getFileContext(
     filename: string,
