+++ packages/rollup-plugin/lib/index.ts
@@ -1,4 +1,4 @@
-export * from '../../compiler/lib/core/constants.ts';
+export * from '../../compiler/lib/core/constants';
 export * from './runtime-compiler';
 export * from './definitions-compiler';
 export * from './types';
