+++ packages/vite-plugin/lib/index.ts
@@ -1,0 +1,2 @@
+export { type JayRollupConfig } from 'rollup-plugin-jay';
+export * from './runtime-compiler';
