+++ packages/vite-plugin/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "jay-dev-environment/library/tsconfig.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "paths": {
+      "jay-compiler": ["../compiler/lib"],
+      "rollup-plugin-jay": ["../rollup-plugin-jay/lib"]
+    }
+  }
+}
